+++ 1D/gid/input_output.hpp
@@ -3,7 +3,7 @@
 void read_input(string filename, Mesh* M){
     string line;
     float k, Q, T_bar, T_hat;
-    short num_nodes, num_elements, num_dirichlet, num_neumann;
+    int num_nodes, num_elements, num_dirichlet, num_neumann;
     ifstream dat_file(filename+".dat");
 
     dat_file >> k >> Q >> T_bar >> T_hat >> num_nodes >> num_elements >> num_dirichlet >> num_neumann;
@@ -16,7 +16,7 @@ void read_input(string filename, Mesh* M){
     dat_file >> line;
 
     for(int i = 0; i < num_nodes; i++){
-        short id;
+        int id;
         float x;
         dat_file >> id >> x;
         M->insert_node(new Node(id,x), i);
@@ -25,7 +25,7 @@ void read_input(string filename, Mesh* M){
     dat_file >> line >> line;
 
     for(int i = 0; i < num_elements; i++){
-        short id, node1_id, node2_id;
+        int id, node1_id, node2_id;
         dat_file >> id >> node1_id >> node2_id;
         M->insert_element( new Element(id, M->get_node(node1_id-1), M->get_node(node2_id-1)) , i);
     }
@@ -33,7 +33,7 @@ void read_input(string filename, Mesh* M){
     dat_file >> line >> line;
 
     for(int i = 0; i < num_dirichlet; i++){
-        short id;
+        int id;
         dat_file >> id;
         M->insert_dirichlet_condition(new Condition(M->get_node(id-1), T_bar), i);
     }
@@ -41,7 +41,7 @@ void read_input(string filename, Mesh* M){
     dat_file >> line >> line;
 
     for(int i = 0; i < num_neumann; i++){
-        short id;
+        int id;
         dat_file >> id;
         M->insert_neumann_condition(new Condition(M->get_node(id-1), T_hat), i);
     }
