+++ 2D/geometry/element.hpp
@@ -1,22 +1,22 @@
 class Element {
     private:
-        short ID;
+        int ID;
         Node* node1;
         Node* node2;
         Node* node3;
 
     public:
-        Element (short identifier, Node* first_node, Node* second_node, Node* third_node){
+        Element (int identifier, Node* first_node, Node* second_node, Node* third_node){
             ID = identifier;
             node1 = first_node;
             node2 = second_node;
             node3 = third_node;
         }
 
-        void set_ID(short identifier){
+        void set_ID(int identifier){
             ID = identifier;
         }
-        short get_ID(){
+        int get_ID(){
             return ID;
         }
